Cast HiddenServiceExportInstanceID to uint16_t before printf.

+++ src/core/or/connection_edge.c
@@ -902,7 +902,7 @@ export_hs_client_circuit_id(edge_connection_t *edge_conn,
   tor_asprintf(&buf, "PROXY TCP6 %s:%s:%04x:%04x:%04x %s %d %d\r\n",
                src_ipv6_prefix,
                rp_identity_buf ? rp_identity_buf : "0000:0000:0000",
-               get_options()->HiddenServiceExportInstanceID,
+               (uint16_t) (get_options()->HiddenServiceExportInstanceID),
                (uint16_t) (gid >> 16), (uint16_t) (gid & 0x0000ffff),
                dst_ipv6, src_port, dst_port);
